move generate individual to its own file
