+++ bliss/datasets/generate_galaxies.py
@@ -1,5 +1,3 @@
-from typing import Optional
-
 import galsim
 import numpy as np
 import torch
@@ -14,7 +12,6 @@ from bliss.datasets.lsst import (
     GALAXY_DENSITY,
     PIXEL_SCALE,
     STAR_DENSITY,
-    convert_mag_to_flux,
     get_default_lsst_background,
 )
 from bliss.datasets.padded_tiles import render_padded_image
@@ -27,31 +24,6 @@ from bliss.datasets.utils import (
     sample_star_fluxes,
     sample_uniform,
 )
-
-
-def generate_individual_dataset(
-    n_samples: int, catsim_table: Table, psf: galsim.GSObject, slen: int = 53, replace: bool = True
-):
-    """Like the function below but it only generates individual galaxies, so much faster to run."""
-
-    background = get_constant_background(get_default_lsst_background(), (n_samples, 1, slen, slen))
-    params, ids = sample_galaxy_params(catsim_table, n_samples, n_samples, replace=replace)
-    assert params.shape == (n_samples, 11)
-    gals = torch.zeros((n_samples, 1, slen, slen))
-    for ii in tqdm(range(n_samples)):
-        gal = render_one_galaxy(params[ii], psf, slen, offset=None)
-        gals[ii] = gal
-
-    # add noise
-    noisy = add_noise_and_background(gals, background)
-
-    return {
-        "images": noisy,
-        "background": background,
-        "noiseless": gals,
-        "galaxy_params": params,
-        "indices": ids,
-    }
 
 
 def generate_dataset(
